backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.8.2
groq==0.11.0
requests==2.32.3
